Project Strawberry/requirements.txt
pandas==1.4.4
Pillow==8.2.0
scikit_image==0.19.3
torch==1.10.0+cu102
torchvision==0.11.0+cu102